writing_agent/requirements.txt
google-adk
google-genai
python-dotenv
textblob